:bug: [Bug] 매너온도 Long->double 타입 변경

+++ src/main/java/samryong/domain/member/dto/MemberDTO.java
@@ -22,7 +22,7 @@ public class MemberDTO {
 
         private String profileImage;
 
-        private Long mannerRate;
+        private double mannerRate;
 
         private LocationResponseDTO location;
